+++ modules/fireback/ReactiveSearchAction.go
@@ -17,8 +17,8 @@ func CreateReactiveSearchHanlder(app *FirebackApp) func(
 	return func(
 		session ReactiveSearchActionSession,
 	) (chan []byte, error) {
-		query := ExtractQueryDslFromGinContext(session.Ctx)
-		query.RawSocketConnection = session.Socket
+		query := ExtractQueryDslFromGinContext(session.GinCtx())
+		query.RawSocketConnection = session.GetSocket()
 		resultChan := make(chan *ReactiveSearchResultDto)
 
 		go func() {
