+++ modules/fireback/ReactiveSearchActionGin.dyno.go
@@ -2,10 +2,8 @@ package fireback
 
 import (
 	"crypto/tls"
-	"fmt"
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
-	"github.com/torabian/emi/emigo"
 	"net/http"
 	"net/url"
 	"unicode/utf8"
@@ -14,133 +12,31 @@ import (
 /**
 * Action to communicate with the action ReactiveSearchAction
  */
-func ReactiveSearchActionMeta() struct {
-	Name        string
-	URL         string
-	Method      string
-	CliName     string
-	Description string
-} {
-	return struct {
-		Name        string
-		URL         string
-		Method      string
-		CliName     string
-		Description string
-	}{
-		Name:        "ReactiveSearchAction",
-		URL:         "/reactive-search",
-		Method:      "REACTIVE",
-		CliName:     "",
-		Description: "Reactive search is a general purpose search mechanism for different modules, and could be used in mobile apps or front-end to quickly search for a entity.",
-	}
-}
-
-/**
- * Query parameters for ReactiveSearchAction
- */
-// Query wrapper with private fields
-type ReactiveSearchActionQuery struct {
-	values url.Values
-	mapped map[string]interface{}
-	// Typesafe fields
-}
-
-func ReactiveSearchActionQueryFromString(rawQuery string) ReactiveSearchActionQuery {
-	v := ReactiveSearchActionQuery{}
-	values, _ := url.ParseQuery(rawQuery)
-	mapped := map[string]interface{}{}
-	if result, err := emigo.UnmarshalQs(rawQuery); err == nil {
-		mapped = result
-	}
-	decoder, err := emigo.NewDecoder(&emigo.DecoderConfig{
-		TagName:          "json", // reuse json tags
-		WeaklyTypedInput: true,   // "1" -> int, "true" -> bool
-		Result:           &v,
-	})
-	if err == nil {
-		_ = decoder.Decode(mapped)
-	}
-	v.values = values
-	v.mapped = mapped
-	return v
-}
-func ReactiveSearchActionQueryFromGin(c *gin.Context) ReactiveSearchActionQuery {
-	return ReactiveSearchActionQueryFromString(c.Request.URL.RawQuery)
-}
-func ReactiveSearchActionQueryFromHttp(r *http.Request) ReactiveSearchActionQuery {
-	return ReactiveSearchActionQueryFromString(r.URL.RawQuery)
-}
-func (q ReactiveSearchActionQuery) Values() url.Values {
-	return q.values
-}
-func (q ReactiveSearchActionQuery) Mapped() map[string]interface{} {
-	return q.mapped
-}
-func (q *ReactiveSearchActionQuery) SetValues(v url.Values) {
-	q.values = v
-}
-func (q *ReactiveSearchActionQuery) SetMapped(m map[string]interface{}) {
-	q.mapped = m
-}
-
 // WebSocket upgrader
 var upgraderReactiveSearchAction = websocket.Upgrader{
 	CheckOrigin: func(r *http.Request) bool { return true },
 }
 
-type ReactiveSearchActionMessage struct {
-	Raw         []byte
-	Conn        *websocket.Conn
-	MessageType int
-	Error       error
+func (x *ReactiveSearchActionMessage) Connection() *websocket.Conn {
+	return x.Conn.(*websocket.Conn)
 }
-
-// Developer handler type
-type ReactiveSearchActionHandler func(msg ReactiveSearchActionMessage) error
+func (x *ReactiveSearchActionClientSession) Connection() *websocket.Conn {
+	return x.Socket.(*websocket.Conn)
+}
+func (x *ReactiveSearchActionSession) GinCtx() *gin.Context {
+	return x.Ctx.(*gin.Context)
+}
+func (x *ReactiveSearchActionSession) GetSocket() *websocket.Conn {
+	return x.Socket.(*websocket.Conn)
+}
 
 // Generated handler
-func ReactiveSearchAction(r *gin.Engine, handler ReactiveSearchActionHandler) {
+func ReactiveSearchActionGin(r *gin.Engine, factory func(
+	session ReactiveSearchActionSession,
+) (chan []byte, error)) {
 	meta := ReactiveSearchActionMeta()
-	r.GET(meta.URL, func(c *gin.Context) {
-		ws, err := upgraderReactiveSearchAction.Upgrade(c.Writer, c.Request, nil)
-		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "cannot upgrade websocket"})
-			return
-		}
-		defer ws.Close()
-		for {
-			mt, raw, err := ws.ReadMessage()
-			msg := ReactiveSearchActionMessage{
-				Conn:        ws,
-				Raw:         raw,
-				Error:       err,
-				MessageType: mt,
-			}
-			// Provide raw message to developer handler
-			if err := handler(msg); err != nil {
-				errMsg := fmt.Sprintf("handler error: %v", err)
-				if writeErr := ws.WriteMessage(mt, []byte(errMsg)); writeErr != nil {
-					break
-				}
-			}
-		}
-	})
+	r.GET(meta.URL, ReactiveSearchActionReactiveHandler(factory))
 }
-
-type ReactiveSearchActionSession struct {
-	Ctx         *gin.Context
-	Socket      *websocket.Conn
-	Done        chan bool
-	Read        chan ReactiveSearchActionReadChan
-	QueryParams ReactiveSearchActionQuery
-}
-type ReactiveSearchActionHandlerDuplex func(*ReactiveSearchActionSession)
-type ReactiveSearchActionReadChan struct {
-	Data  []byte
-	Error error
-}
-
 func ReactiveSearchActionReactiveHandler(factory func(
 	session ReactiveSearchActionSession,
 ) (chan []byte, error)) gin.HandlerFunc {
@@ -159,7 +55,7 @@ func ReactiveSearchActionReactiveHandler(factory func(
 			Done:   done,
 			Read:   read,
 		}
-		session.QueryParams = ReactiveSearchActionQueryFromGin(ctx)
+		session.QueryParams = ReactiveSearchActionQueryFromHttp(ctx.Request)
 		write, err := factory(session)
 		if err != nil {
 			c.WriteMessage(websocket.TextMessage, []byte(err.Error()))
@@ -203,18 +99,6 @@ func ReactiveSearchActionReactiveHandler(factory func(
 			}
 		}()
 	}
-}
-
-// ReactiveSearchActionClientSession is the client-side mirror of
-// ReactiveSearchActionSession. Receive frames on Read, send frames on Write,
-// and close Write (or send on Done) to tear the connection down. Done also
-// fires when the server closes or the socket errors, so the caller can use it
-// as a single disconnect signal.
-type ReactiveSearchActionClientSession struct {
-	Socket *websocket.Conn
-	Done   chan bool
-	Read   chan ReactiveSearchActionReadChan
-	Write  chan []byte
 }
 
 // ReactiveSearchActionClientOptions configures a client dial. All fields are
